+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - User : Add a search filter to see the active users.
 - Add support for multiple demo data config per app. 
 - Configurator creator : Add demo data.
+- Add data backup and restore.
 
 ## Improvements
 - Stock rules : change column name of code and name to product code and product name.